+++ src/main/java/com/atlauncher/gui/tabs/SettingsTab.java
@@ -28,6 +28,7 @@ import javax.swing.JTabbedPane;
 import org.mini2Dx.gettext.GetText;
 
 import com.atlauncher.App;
+import com.atlauncher.builders.HTMLBuilder;
 import com.atlauncher.evnt.listener.RelocalizationListener;
 import com.atlauncher.evnt.manager.RelocalizationManager;
 import com.atlauncher.evnt.manager.SettingsManager;
@@ -39,7 +40,9 @@ import com.atlauncher.gui.tabs.settings.JavaSettingsTab;
 import com.atlauncher.gui.tabs.settings.LoggingSettingsTab;
 import com.atlauncher.gui.tabs.settings.ModsSettingsTab;
 import com.atlauncher.gui.tabs.settings.NetworkSettingsTab;
+import com.atlauncher.managers.DialogManager;
 import com.atlauncher.network.Analytics;
+import com.atlauncher.utils.OS;
 import com.formdev.flatlaf.FlatLaf;
 
 @SuppressWarnings("serial")
@@ -80,6 +83,7 @@ public class SettingsTab extends JPanel implements Tab, RelocalizationListener {
                     && networkSettingsTab.canConnectWithProxy()) {
                 boolean reloadTheme = generalSettingsTab.needToReloadTheme();
                 boolean themeChanged = generalSettingsTab.themeChanged();
+                boolean languageChanged = generalSettingsTab.languageChanged();
                 boolean reloadInstancesPanel = generalSettingsTab.needToReloadInstancesPanel();
                 generalSettingsTab.save();
                 modsSettingsTab.save();
@@ -95,6 +99,19 @@ public class SettingsTab extends JPanel implements Tab, RelocalizationListener {
                 }
                 if (themeChanged) {
                     Analytics.sendEvent(App.THEME.getName(), "ChangeTheme", "Launcher");
+                }
+                if (languageChanged) {
+                    int ret = DialogManager.yesNoDialog().setType(DialogManager.INFO)
+                            .setTitle(GetText.tr("Language Changed. Restart?"))
+                            .setContent(new HTMLBuilder().center().text(GetText.tr(
+                                    "You've changed the language. For best results, a restart of the launcher is recommended.<br/><br/>Restart Now?"))
+                                    .build())
+                            .show();
+
+                    if (ret == 0) {
+                        OS.restartLauncher();
+                        return;
+                    }
                 }
                 if (reloadTheme) {
                     App.loadTheme(App.settings.theme);