+++ src/main/java/com/atlauncher/gui/tabs/settings/GeneralSettingsTab.java
@@ -525,6 +525,10 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
         return !((ComboItem<String>) theme.getSelectedItem()).getValue().equalsIgnoreCase(App.settings.theme);
     }
 
+    public boolean languageChanged() {
+        return !((String) language.getSelectedItem()).equalsIgnoreCase(App.settings.language);
+    }
+
     public boolean needToReloadInstancesPanel() {
         return !(((ComboItem<String>) instanceTitleFormat.getSelectedItem()).getValue())
                 .equals(App.settings.instanceTitleFormat);