+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 ## 3.4.19.2
 
 ### New Features
+- Add prompt to restart launcher when changing language
 
 ### Fixes
 - Issue with NPEs when checking for CurseForge mod updates [#608]